requirements.txt
pandas
yfinance
jupyter
requests
anthropic
python-dotenv